Revised API for VotingSystems.
Ensured all VotingSystem subclasses have a default constructor so that
instances can be created via reflection during election configuration.

UserInterface now has a constructor that attaches the UI to an 
Election.

VotingSystem now creates a Queue for VoterActions and has a new
setElection feature to permit the attachment of an election to a 
VotingSystem after construction via reflection.